LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/relations/css-styling/padding-border-margin/margin-001-expected.txt


FAIL Margin properties on mrow assert_approx_equals: top margin expected 40 +/- 1 but got 80
FAIL Margin properties on mrow (rtl) assert_approx_equals: top margin expected 40 +/- 1 but got 80
FAIL Margin properties on mrow (shorthand) assert_approx_equals: top margin expected 20 +/- 1 but got 40
FAIL Margin properties on mrow (logical) assert_approx_equals: top margin expected 40 +/- 1 but got 80
FAIL Margin properties on mrow (logical, rtl) assert_approx_equals: top margin expected 40 +/- 1 but got 80
FAIL Margin properties on mrow (logical, shorthand) assert_approx_equals: top margin expected 30 +/- 1 but got 60
